Split Definition & Literal into their own module
